LayoutTests/fast/events/popup-allowed-from-gesture-initiated-event.html

<html> 
    <head>
        <link rel="stylesheet" href="../js/resources/js-test-style.css">
        <script src="../js/resources/js-test-pre.js"></script>
    </head>
    <body>
        <button id="button1" onclick="button1()">Click Here</button>
        <button id="button2" onclick="button2()">Click Here Too</button>
        <button id="test" onclick="popup()" style="display:none"></button>
        <div id="console"></div>
        <script> 
            var win;
            function popup() {
                win = window.open("about:blank", "blank");
                shouldBeNonNull("win");
            }
 
            function button1() {
                document.getElementById("test").click();
            }

            function button2() {
                var clickEvent = document.createEvent("MouseEvents");
                clickEvent.initMouseEvent("click", true, true, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
                document.getElementById("test").dispatchEvent(clickEvent);
            }

            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.setCanOpenWindows();
                layoutTestController.setPopupBlockingEnabled(true);
                layoutTestController.setCloseRemainingWindowsWhenComplete(true);
                layoutTestController.waitUntilDone();

                if (window.eventSender) {
                    eventSender.mouseMoveTo(10, 10);
                    eventSender.mouseDown();
                    eventSender.mouseUp();
                    eventSender.mouseMoveTo(100, 10);
                    eventSender.mouseDown();
                    eventSender.mouseUp();
                }

                layoutTestController.notifyDone();
            }
        </script> 
    </body> 
</html>
