LayoutTests/fast/events/popup-allowed-from-gesture-initiated-event-expected.txt

Click Here  Click Here Too
PASS win is non-null.
PASS win is non-null.
